+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Example `datasource-health-check`: Without `--uid` option, scan the whole
   Grafana instance
 * Data source health check: Be graceful on network read timeouts
+* Data source health check: Improve acceptance criteria when probing Prometheus
 
 
 ## 2.3.0 (2022-05-26)